test: refactor functional test structure
